refactor: move checkCorrectData to a separate file
