+++ app/src/espresso/java/xyz/lbres/trickcalculator/ui/history/testRandomness2.kt
@@ -4,16 +4,16 @@ import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withText
-import xyz.lbres.kotlinutils.general.ternaryIf
 import xyz.lbres.trickcalculator.testutils.closeFragment
 import xyz.lbres.trickcalculator.testutils.openHistoryFragment
-import xyz.lbres.trickcalculator.testutils.repeatUntil
 import xyz.lbres.trickcalculator.testutils.toggleShuffleOperators
 import xyz.lbres.trickcalculator.ui.calculator.clearText
 import xyz.lbres.trickcalculator.ui.calculator.equals
 import xyz.lbres.trickcalculator.ui.calculator.typeText
 
 fun testRandomness2() {
+    val errorMessage = "History items and pairs should be shuffled in history randomness 2."
+
     setHistoryRandomness(2)
     toggleShuffleOperators()
     openHistoryFragment()
@@ -28,73 +28,44 @@ fun testRandomness2() {
     typeText("400/5")
     equals()
     computeHistory.add(Pair("400/5", "80"))
-    checkCorrectData(computeHistory)
+    checkCorrectData(computeHistory, 2, errorMessage)
 
     // several elements
     clearText()
     typeText("15-2.5")
     equals()
     computeHistory.add(Pair("15-2.5", "12.5"))
-    checkCorrectData(computeHistory)
+    checkCorrectData(computeHistory, 2, errorMessage)
 
     clearText()
     typeText("(3-4)(5+2)")
     equals()
     computeHistory.add(Pair("(3-4)(5+2)", "-7"))
-    checkCorrectData(computeHistory)
+    checkCorrectData(computeHistory, 2, errorMessage)
 
     // previously computed
     typeText("+11")
     equals()
     computeHistory.add(Pair("-7+11", "4"))
-    checkCorrectData(computeHistory)
+    checkCorrectData(computeHistory, 2, errorMessage)
 
     // duplicate element
     clearText()
     typeText("(3-4)(5+2)")
     equals()
     computeHistory.add(Pair("(3-4)(5+2)", "-7"))
-    checkCorrectData(computeHistory)
+    checkCorrectData(computeHistory, 2, errorMessage)
 
     // error
     clearText()
     typeText("+")
     equals()
     computeHistory.add(Pair("+", "Syntax error"))
-    checkCorrectData(computeHistory)
+    checkCorrectData(computeHistory, 2, errorMessage)
 
     clearText()
     typeText("2^0.5")
     equals()
     computeHistory.add(Pair("2^0.5", "Exponents must be whole numbers"))
-    checkCorrectData(computeHistory)
-}
-
-/**
- * Check that all the expected information is displayed in the history.
- * Verifies that all items are displayed, the order of items is shuffled, and the pairs of items are shuffled.
- *
- * @param computeHistory [TestHistory]: list of items in history
- */
-private fun checkCorrectData(computeHistory: TestHistory) {
-    var shuffled = false
-    val historySize = computeHistory.size
-    val checker = HistoryChecker(computeHistory)
-
-    // additional repeats for 2 items due to occasional failures
-    val repeatCount = ternaryIf(historySize == 2, 10, 5)
-
-    repeatUntil(repeatCount, { shuffled }) {
-        openHistoryFragment()
-
-        // check that all items are displayed
-        checker.checkDisplayed(2)
-        shuffled = shuffled || checker.checkShuffled(2)
-
-        closeFragment()
-    }
-
-    if (historySize > 1 && !shuffled) {
-        throw AssertionError("History items and pairs should be shuffled in history randomness 2. History: $computeHistory")
-    }
+    checkCorrectData(computeHistory, 2, errorMessage)
 }
